ai_service/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
python-dotenv
openai
python-multipart
pytest
pytest-asyncio

qdrant-client
pdfminer.six